Updated to use latest Docker v1.11.0 and Volume API
This commit updates the volume API implementation to leverage
the Docker v1.11.0. There was an addition of Get() and List()
methods in prior v1.10. The Get() method now replaces the
call that implicitly created volumes. The Create() method
was used for Path() and other methods prior because there
was no Get() method that returned a volume object.